requirements.txt
Flask
pyjwt==1.5.3
mongoengine
passlib
pyotp
bcrypt
pymongo[srv]